refactor(define-router): split config utils
